Updated symbols

+++ frontend/packages/naukluft-data-backend/sql/sections/symbols.sql
@@ -2,8 +2,8 @@ SELECT
   id,
   section::text section_id,
   symbol,
-  start_height,
-  end_height,
+  start_height::numeric,
+  end_height::numeric,
   coalesce(symbol_min_zoom, 0) symbol_min_zoom,
   (
     start_height+coalesce(
